# RadiCorder/openapi/openapi.local.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RadiCorder | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5011/"
    }
  ],
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "NotificationsApi"
        ],
        "summary": "お知らせ一覧を取得する",
        "operationId": "ApiNotificationList",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfNotificationListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reserves/keywords": {
      "get": {
        "tags": [
          "ReservesApi"
        ],
        "summary": "キーワード予約一覧を取得する",
        "operationId": "ApiReserveKeywords",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfListOfKeywordReserveEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/reserves/keywords/update": {
      "post": {
        "tags": [
          "ReservesApi"
        ],
        "summary": "キーワード予約を更新する",
        "operationId": "ApiReserveKeywordUpdate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordReserveEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/reserves/keywords/delete": {
      "post": {
        "tags": [
          "ReservesApi"
        ],
        "summary": "キーワード予約を削除する",
        "operationId": "ApiReserveKeywordDelete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/reserves/keywords/switch": {
      "post": {
        "tags": [
          "ReservesApi"
        ],
        "summary": "キーワード予約の有効状態を切り替える",
        "operationId": "ApiReserveKeywordSwitch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/reserves/keywords/reorder": {
      "post": {
        "tags": [
          "ReservesApi"
        ],
        "summary": "キーワード予約の並び順を更新する",
        "operationId": "ApiReserveKeywordReorder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordReserveReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/reserves/programs": {
      "get": {
        "tags": [
          "ReservesApi"
        ],
        "summary": "番組予約一覧を取得する",
        "operationId": "ApiReservePrograms",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfListOfScheduleEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    },
    "/api/reserves/programs/delete": {
      "post": {
        "tags": [
          "ReservesApi"
        ],
        "summary": "番組予約を削除する",
        "operationId": "ApiReserveProgramDelete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfEmptyData"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": [
              "null",
              "string"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseOfEmptyData": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EmptyData"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfListOfKeywordReserveEntry": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/KeywordReserveEntry"
            }
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfListOfScheduleEntry": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ScheduleEntry"
            }
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfNotificationListResponse": {
        "required": [
          "success",
          "data",
          "error",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NotificationListResponse"
              }
            ]
          },
          "error": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiError"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DaysOfWeek": {
        "type": "integer"
      },
...
      "EmptyData": {
        "type": "object"
      },
...
      "KeywordReserveEntry": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Ulid"
          },
          "selectedRadikoStationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "selectedRadiruStationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "keyword": {
            "type": "string"
          },
          "excludedKeyword": {
            "type": "string"
          },
          "recordPath": {
            "type": "string"
          },
          "recordFileName": {
            "type": "string"
          },
          "searchTitleOnly": {
            "type": "boolean"
          },
          "excludeTitleOnly": {
            "type": "boolean"
          },
          "selectedDaysOfWeek": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DaysOfWeek"
            }
          },
          "startTimeString": {
            "type": "string"
          },
          "endTimeString": {
            "type": "string"
          },
          "startTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "endTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "startDelay": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "endDelay": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "sortOrder": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mergeTagBehavior": {
            "$ref": "#/components/schemas/KeywordReserveTagMergeBehavior"
          }
        }
      },
      "KeywordReserveReorderRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "KeywordReserveTagMergeBehavior": {
        "type": "integer"
      },
...
      "NoticeCategory": {
        "type": "integer"
      },
      "NotificationEntry": {
        "type": "object",
        "properties": {
          "logLevel": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/NoticeCategory"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "isRead": {
            "type": "boolean"
          }
        }
      },
...
      "NotificationListResponse": {
        "required": [
          "totalRecords",
          "page",
          "pageSize",
          "recordings"
        ],
        "type": "object",
        "properties": {
          "totalRecords": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "recordings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationEntry"
            }
          }
        }
      },
...
      "RadioServiceKind": {
        "type": "integer"
      },
...
      "RecordingType": {
        "type": "integer"
      },
      "ReserveEntryRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "ReserveType": {
        "type": "integer"
      },
      "ScheduleEntry": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Ulid"
          },
          "serviceKind": {
            "$ref": "#/components/schemas/RadioServiceKind"
          },
          "stationId": {
            "type": "string"
          },
          "stationName": {
            "type": "string"
          },
          "areaId": {
            "type": "string"
          },
          "programId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "performer": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "recordingType": {
            "$ref": "#/components/schemas/RecordingType"
          },
          "reserveType": {
            "$ref": "#/components/schemas/ReserveType"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "matchedKeywordReserveKeywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "plannedTagNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Ulid": { },
...
    }
  },
...
}